requirements.txt
openai>=0.27.0
streamlit>=1.22.0
python-dotenv>=0.19.0
streamlit-authenticator==0.1.5
streamlit-javascript>=0.1.0
chardet>=4.0.0
openpyxl>=3.0.0
python-docx>=0.8.10
requests>=2.20.0